+++ src/lib.rs
@@ -25,7 +25,6 @@ pub mod metrics;
 pub mod multipart;
 pub mod rate_limiter;
 pub mod replication;
-#[cfg(feature = "s3s-adapter")]
 pub mod s3_adapter_s3s;
 pub mod secret;
 pub mod security;